+++ ProjectEulerSolutions/ConceptualMath/Numbers/INumber.cs
@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Numerics;
 using System.Text;
 
 namespace ConceptualMath.Numbers
 {
     public interface INumber
     {
-        int Value { get; }
+        BigInteger Value { get; }
     }
 }
